Use Vector/Matrix classes, update to .net 4.6, DAE fixes, more UI work
The classes from System.Vector are now used instead of the custom ones on Math3D. This was done to avoid reinventing the well. Support for the used classes were added on the serializer. Some issues were fixed on the DAE exporter, one of which made 3dsmax/Maya crash when trying to open dae files with vertex colors. Work to unclutter the UI and keep only UI Control code on the main form continues. Some small UI issues were fixed aswell.